chore: introduce a `Categoria` model
Right now a `Producto` has a prop named `Categoria`,
the idea is simple, this prop has a string value representing
the category of the product.

The problem with the current schema is that when records
of products were added, the repetition of categories names
was obvious. So, if in the future the name of a category
has to change, the change has to be made in multiple places.
And thus, more work.

This commit introduces and documents a new model named `Categoria`.
with this model in place, changing a category name is as simple as
changing it in one place.